chatterbox_requirements.txt
chatterbox-tts
torchaudio
fastapi
uvicorn
python-multipart
python-dotenv
